+++ pizzastore/pizzastore-api/infra-web/src/main/java/com/nortal/pizzastore/auth/PasswordEncoderImpl.java
@@ -2,6 +2,7 @@ package com.nortal.pizzastore.auth;
 
 import com.nortal.pizzastore.usecase.registeruser.Encoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -9,17 +10,13 @@ class PasswordEncoderImpl implements Encoder, PasswordEncoder {
 
   @Override
   public String encode(CharSequence rawPassword) {
-
-    // TODO! Proper password encoding
-
-    return rawPassword.toString();
+    BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
+    return encoder.encode(rawPassword);
   }
 
   @Override
   public boolean matches(CharSequence rawPassword, String encodedPassword) {
-
-    // TODO! Proper password encoding
-
-    return rawPassword.toString().equals(encodedPassword);
+    BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
+    return encoder.matches(rawPassword, encodedPassword);
   }
 }